requirements.txt
GPUtil==1.4.0
future==0.17.1
ImageHash==4.0
joblib==0.13.2
ngt==1.7.6
numpy==1.16.4
opencv-python==4.1.0.25
orderedset==2.0.1
pathlib==1.0.1
pathos==0.2.4
Pillow==6.1.0
scipy==1.3.0
six==1.12.0
termcolor==1.1.0
tqdm==4.32.2
webcolors==1.9.1
pyheif==0.6.0
whatimage==0.0.3